Move tree binding stuff into one file as well
